+++ cg/services/fastq_file_service/fastq_file_service.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 
-from cg.services.fastq_service.utils import concatenate_forward_reads, concatenate_reverse_reads
+from cg.services.fastq_file_service.utils import concatenate_forward_reads, concatenate_reverse_reads
 
 
-class FastqService:
+class FastqFileService:
 
     def concatenate(self, fastq_directory: Path, forward_output: Path, reverse_output: Path):
         concatenate_forward_reads(directory=fastq_directory, output_file=forward_output)
